Move `NonNullByDefault` to util module
