+++ standalone/src/main/java/dev/gihwan/tollgate/standalone/package-info.java
@@ -1,7 +1,7 @@
 /*
  * MIT License
  *
- * Copyright (c) 2020 Gihwan Kim
+ * Copyright (c) 2020 - 2021 Gihwan Kim
  *
  * Permission is hereby granted, free of charge, to any person obtaining a copy
  * of this software and associated documentation files (the "Software"), to deal
@@ -25,4 +25,4 @@
 @NonNullByDefault
 package dev.gihwan.tollgate.standalone;
 
-import dev.gihwan.tollgate.core.annotation.NonNullByDefault;
+import dev.gihwan.tollgate.util.NonNullByDefault;
